Version dependencies with Go modules
The project dependency versions weren't being tracked anywhere, as far
as I could tell. That can make for nasty surprises if any of the
upstream packages make breaking changes to their APIs. We can avoid this
by explicitly fixing known versions of the dependencies that work with
the project, and updating them explicitly when necessary.

There are a number of tools that can do this in the Go ecosystem. As of
this year, the winner seems to be Go modules. As the Go project has
adopted it, it at least should have a chance of not being dropped on the
floor in the near future.

To run inside GOPATH, this still needs the environment variable

    GO111MODULE=on

to be set.