Add option to log errors to STDERR in metha-sync
Logrus logs all messages to STDERR by default. If we direct its messages
elsewhere, it does not have a built-in way of logging only errors to
STDERR. This is unfortunate if our program is being run by a runner that
believes that messages on STDERR are something it should tell us about.
In that case, we either get no messages in case of errors, or get
messages even on successful runs where there were no problems.

This happens, for example, if we tell cron to run metha-sync for us.

This patch proposes a solution: If we give the new -log-errors-to-stderr
option to metha-sync, it will add a hook to logrus that copies messages
with level WARN and above to STDERR. The whole log message stream will
either be written to the file given by -log, ignored if -q was given, or
written to STDOUT otherwise so we won't double-write warnings and
errors.

If a user wants to log errors to STDERR and the log stream to STDOUT,
they'll nonetheless get a single warning message to STDERR saying that
they haven't specified -q or -log, and that we're deviating from the
default behavior. They can avoid this by giving the value /dev/stdout to
-log, which seems reasonably portable:

    https://unix.stackexchange.com/questions/36403/how-portable-are-dev-stdin-dev-stdout-and-dev-stderr/278368#278368